api_approach/requirements.txt
pdfplumber==0.10.3
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==0.21.0
selenium==4.15.2
